+++ horovod/ray/__init__.py
@@ -1,1 +1,4 @@
 from .runner import RayExecutor, BaseHorovodWorker
+from .elastic import ElasticRayExecutor
+
+__all__ = ["RayExecutor", "BaseHorovodWorker", "ElasticRayExecutor"]