+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Added Elastic Ray integration. ([#2291](https://github.com/horovod/horovod/pull/2291))
+
 ### Changed
 
 ### Deprecated